requirements.txt
mss
pyautogui
pygetwindow
pyperclip
requests
pyyaml>=6.0
openai>=1.0.0
